# services/api-gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api-gateway",
    "description": "Single authenticated ingress over the platform's internal services. Services and AI agents call it with a bearer token to read catalog products, offers and price history, and to curate matches — one consistent, scoped, rate-limited contract.",
    "contact": {
      "name": "Supplier Aggregation Platform",
      "email": "Info_DIT@erc.ua"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/curation/stats": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Curation stats for the operator dashboard",
        "description": "Aggregate counts for the operator dashboard: queue depth, total canonical products, and decisions by action. Requires scope `matching:curate`.",
        "operationId": "getCurationStats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/curation/queue": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "List the matching curation queue",
        "description": "List supplier→canonical link suggestions awaiting an operator decision, oldest first. Cursor-paginated. Requires scope `matching:curate`.",
        "operationId": "getCurationQueue",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous response's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous response's next_cursor."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max items per page (1-200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max items per page (1-200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/curation/decisions": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "List the curation decision journal",
        "description": "Append-only audit log of curation decisions (confirm / reject / create-new / merge), newest first. Optionally filter to one supplier product's history. Cursor-paginated. Requires scope `matching:curate`.",
        "operationId": "getCurationDecisions",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one supplier product's decision history.",
              "title": "Supplier Product Id"
            },
            "description": "Filter to one supplier product's decision history."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous response's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous response's next_cursor."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max items per page (1-200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max items per page (1-200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/curation/links/{supplier_product_id}/confirm": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Confirm a suggested match",
        "description": "Confirm the suggested canonical for a supplier product; emits matching.link.confirmed downstream. Idempotent. Requires scope `matching:curate`.",
        "operationId": "confirmCurationLink",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID) to confirm.",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID) to confirm."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/curation/links/{supplier_product_id}/reject": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Reject a suggested match",
        "description": "Reject the suggested canonical for a supplier product. Idempotent. Requires scope `matching:curate`.",
        "operationId": "rejectCurationLink",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID) to reject.",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID) to reject."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/curation/links/{supplier_product_id}/create-new": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Create a new canonical from a curation item",
        "description": "Reject the suggested candidate and create a brand-new canonical product from this supplier product, linking it confirmed. Use it when the suggestion is wrong and no existing canonical fits. Requires scope `matching:curate`.",
        "operationId": "createNewCanonical",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID).",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCanonicalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateCanonicalIn": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Canonical title for the new product; prefill from the supplier."
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand",
            "description": "Brand/vendor name, if known."
          },
          "gtin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gtin",
            "description": "Normalized GTIN-14 to assign; must not already belong to another canonical."
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateCanonicalIn",
        "description": "Fields for a brand-new canonical product created from a curation item (mirrors matching)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Stable URI identifying the error type."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary of the error type."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human/LLM-readable explanation with a next step."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "URI of the specific occurrence."
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id",
            "description": "Trace id to correlate with telemetry."
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem+json error body (returned on 4xx/5xx)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service API key or OIDC token, issued/validated by the API gateway."
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This service (behind the API gateway)."
    }
  ],
...
}